ENH: reformulated plotting functions
Moved the plotting functions into new sub-modules.
