+++ pyValEIA/plots/__init__.py
@@ -5,10 +5,8 @@
 """Submodule for plotting data."""
 
 # Import the package modules and top-level classes
-from pyValEIA.plots import Madrigal_NIMO2  # noqa F401
-from pyValEIA.plots import NIMO_Swarm_Map_Plotting  # noqa F401
-from pyValEIA.plots import NIMO_SWARM_single  # noqa F401
-from pyValEIA.plots import offset_codes  # noqa F401
-from pyValEIA.plots import paper_plotting  # noqa F401
-from pyValEIA.plots import swarm_panel_ax  # noqa F401
-from pyValEIA.plots import SwarmPyIRI  # noqa F401
+from pyValEIA.plots import mad_diagnostic_plots  # noqa F401
+from pyValEIA.plots import mad_stat_plots  # noqa F401
+from pyValEIA.plots import swarm_diagnostic_plots  # noqa F401
+from pyValEIA.plots import swarm_stat_plots  # noqa F401
+from pyValEIA.plots import utils  # noqa F401
